requirements.txt
customtkinter
selenium
minecraft-launcher-lib
wget
requests
Pillow
imagehash
pywinstyles
psutil